+++ tests/test_segment.py
@@ -7,32 +7,40 @@ from dragodis import SegmentType, SegmentPermission
 def test_basic(disassembler):
     if disassembler.name.lower() == "ida":
         expected_segments = [
-            (0x00401000, 0x00409C00, ".text", SegmentPermission.read | SegmentPermission.execute),
-            (0x0040A000, 0x0040A110, ".idata", SegmentPermission.read),
-            (0x0040A110, 0x0040BE00, ".rdata", SegmentPermission.read),
-            (0x0040C000, 0x0040D200, ".data", SegmentPermission.read | SegmentPermission.write),
+            (0x00401000, 0x0040A000, ".text", True, SegmentPermission.read | SegmentPermission.execute),
+            (0x0040A000, 0x0040A110, ".idata", False, SegmentPermission.read),
+            (0x0040A110, 0x0040C000, ".rdata", True, SegmentPermission.read),
+            (0x0040C000, 0x0040F000, ".data", True, SegmentPermission.read | SegmentPermission.write),
         ]
     elif disassembler.name.lower() == "ghidra":
         expected_segments = [
-            (0x00400000, 0x00400400, "Headers", SegmentPermission.read),
-            (0x00401000, 0x00409C00, ".text", SegmentPermission.read | SegmentPermission.execute),
-            (0x0040A000, 0x0040BE00, ".rdata", SegmentPermission.read),
-            (0x0040C000, 0x0040D200, ".data", SegmentPermission.read | SegmentPermission.write),
+            (0x00400000, 0x00400400, "Headers", True, SegmentPermission.read),
+            (0x00401000, 0x00409C00, ".text", True, SegmentPermission.read | SegmentPermission.execute),
+            (0x0040A000, 0x0040BE00, ".rdata", True, SegmentPermission.read),
+            (0x0040C000, 0x0040D200, ".data", True, SegmentPermission.read | SegmentPermission.write),
+            (0x0040D200, 0x0040ed48, ".data", False, SegmentPermission.read | SegmentPermission.write),
         ]
     else:
         raise NotImplementedError
 
     actual_segments = list(disassembler.segments)
+    print(actual_segments)
     assert len(actual_segments) == len(expected_segments)
     for segment, expected_segments in zip(actual_segments, expected_segments):
-        start, end, name, permissions = expected_segments
+        start, end, name, initialized, permissions = expected_segments
         assert segment
         assert hex(segment.start) == hex(start)
         assert hex(segment.end) == hex(end)
+        assert segment.name == name
+        assert segment.initialized == initialized
         assert segment.permissions == permissions
         assert start in segment
+        assert start + 1 in segment
 
         # Test direct retrieval
+        # Doesn't work for the Ghidra segment duplicate.
+        if name == ".data" and not initialized:
+            continue
         segment2 = disassembler.get_segment(name)
         assert segment2 == segment
         segment3 = disassembler.get_segment(start)